Refactor code and add Mat as a property of a CardPile
